change in app tips, upgrade them from byte to int

+++ windows-client/DbUtils/MiscDBUtil.cs
@@ -139,8 +139,8 @@ namespace windows_client.DbUtils
             #endregion
             #region RESET IN APP TIPS
             App.appSettings[App.CHAT_THREAD_COUNT_KEY] = 0;
-            App.appSettings[App.TIP_MARKED_KEY] = (byte)0;
-            App.WriteToIsoStorageSettings(App.TIP_SHOW_KEY, (byte)0); // to keep a track of current showing keys
+            App.appSettings[App.TIP_MARKED_KEY] = 0;
+            App.WriteToIsoStorageSettings(App.TIP_SHOW_KEY, 0); // to keep a track of current showing keys
             App.ViewModel.LoadToolTipsDict();
             #endregion
         }
